Completed migration of Binary Search
