+++ coollime/src/test/java/coollime/binarysearch/kclosestinsortedarray/KClosestInSortedArrayTest.java
@@ -1,0 +1,41 @@
+package coollime.binarysearch.kclosestinsortedarray;
+
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+class KClosestInSortedArrayTest {
+  private static Solution sut;
+
+  @BeforeAll
+  static void setInstance() {
+    sut = new Solution();
+  }
+
+  @AfterEach
+  void tearDown() {
+  }
+
+  @Test
+  void shouldNotFindTheResultWhenTheArrayIsNull() {
+    int[] result = sut.kClosest(null, 1, 1);
+    assertArrayEquals(new int[] {}, result);
+  }
+
+  @Test
+  void shouldNotFindTheResultWhenTheArrayIsEmpty() {
+    int[] result = sut.kClosest(new int[] {}, 1, 1);
+    assertArrayEquals(new int[] {}, result);
+  }
+
+  @Test
+  void shouldFindTheResult() {
+    int[] result = sut.kClosest(new int[] { 1, 2, 3 }, 2, 3);
+    assertArrayEquals(new int[] { 2, 1, 3 }, result);
+
+    result = sut.kClosest(new int[] { 1, 4, 6, 8 }, 3, 3);
+    assertArrayEquals(new int[] { 4, 1, 6 }, result);
+  }
+}
