+++ docs/binarysearch/README.md
@@ -1,0 +1,10 @@
+## Binary Search
+
+1.  [Classical Binary Search](ClassicalBinarySearch)
+2.  [Search in Sorted Matrix I](SearchInSortedMatrixI)
+3.  [Closest in Sorted Array](ClosestInSortedArray)
+4.  [First Occurrence](FirstOccurrence)
+5.  [Last Occurrence](LastOccurrence)
+6.  [K Closest in Sorted Array](KClosestInSortedArray)
+7.  [Search in Unknown Sized Sorted Array](SearchInUnknownSizedSortedArray)
+8.  [Find Minimum in Rotated Sorted Array](FindMinimumInRotatedSortedArray)
