+++ docs/binarysearch/search_sorted_matrix.png
