--- src/B/BinarySearch/README.md
@@ -1,9 +1,0 @@
-## Binary Search
-1.  [Classical Binary Search](Easy/ClassicalBinarySearch)
-2.  [Search in Sorted Matrix I](Medium/SearchInSortedMatrixI)
-3.  [Closest in Sorted Array](Medium/ClosestInSortedArray)
-4.  [First Occurrence](Medium/FirstOccurrence)
-5.  [Last Occurrence](Medium/LastOccurrence)
-6.  [K Closest in Sorted Array](Medium/KClosestInSortedArray)
-7.  [Search in Unknown Sized Sorted Array](Medium/SearchInUnknownSizedSortedArray)
-8.  [Find Minimum in Rotated Sorted Array](Medium/FindMinimumInRotatedSortedArray)
